+++ documentation/docs/release-notes.md
@@ -7,6 +7,7 @@ title: Release notes&#58;
 
 - Support SAML2 `Scoping` in authentication requests.
 - `WebContext` is now able to provide the request url directly.
+- Fix SAML2 response attributes resolution when using ADFS as IdP.
 
 **v5.1.1**: